refactor: split Jest configs
